requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.3.1
colorama==0.4.6
Deprecated==1.3.1
ecdsa==0.19.1
Flask==3.1.3
Flask-Caching==2.3.1
Flask-Limiter==3.8.0
flask-marshmallow==1.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
mysql-connector-python==9.6.0
ordered-set==4.1.0
packaging==26.0
pyasn1==0.6.2
Pygments==2.19.2
python-jose==3.3.0
rich==13.9.4
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.6
wrapt==2.1.1
